requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
cryptography==35.0.0
idna==3.2
multidict==5.2.0
pycparser==2.20
PyMySQL==1.0.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
SimpleQIWI==0.7
six==1.16.0
typing-extensions==3.10.0.2
tzlocal==2.1
urllib3==1.26.7
uuid==1.30
yarl==1.7.0
